pyproject.toml
[tool.poetry]
name = "djangopdfchat"
version = "0.1.0"
description = ""
authors = ["Serban Marti <marti.serban@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^4.2.13"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
psycopg = "^3.2.1"
gunicorn = "^22.0.0"
djoser = "^2.2.3"
pdf2image = "^1.17.0"
openai = "^1.35.10"
drf-spectacular = "^0.27.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-django = "^4.8.0"
pylint = "^3.2.5"
pylint-django = "^2.5.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
    "."
]